+++ crates/dslab-dag/README.md
@@ -16,8 +16,9 @@ use std::cell::RefCell;
 
 use dslab_dag::dag::DAG;
 use dslab_dag::dag_simulation::DagSimulation;
+use dslab_dag::data_item::DataTransferMode;
 use dslab_dag::network::load_network;
-use dslab_dag::runner::{Config, DataTransferMode};
+use dslab_dag::runner::Config;
 use dslab_dag::schedulers::simple_scheduler::SimpleScheduler;
 
 // load network model configuration
