+++ crates/dslab-dag/src/schedulers/mod.rs
@@ -1,5 +1,6 @@
 //! Scheduler implementations.
 
+mod common;
 pub mod heft;
 pub mod lookahead;
 pub mod portfolio_scheduler;
