+++ crates/dslab-dag/src/schedulers/portfolio_scheduler.rs
@@ -4,7 +4,8 @@ use dslab_core::context::SimulationContext;
 use dslab_network::network::Network;
 
 use crate::dag::DAG;
-use crate::runner::{Config, DataTransferMode};
+use crate::data_item::DataTransferMode;
+use crate::runner::Config;
 use crate::scheduler::{Action, Scheduler};
 use crate::task::*;
