+++ examples/dag-benchmark/src/main.rs
@@ -7,8 +7,9 @@ use sugars::{rc, refcell};
 
 use dslab_dag::dag::DAG;
 use dslab_dag::dag_simulation::DagSimulation;
+use dslab_dag::data_item::DataTransferMode;
 use dslab_dag::network::load_network;
-use dslab_dag::runner::{Config, DataTransferMode};
+use dslab_dag::runner::Config;
 use dslab_dag::schedulers::simple_scheduler::SimpleScheduler;
 
 fn main() {
